backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
beanie==1.27.0
motor==3.6.0
pydantic-settings==2.7.1
bcrypt==5.0.0
python-jose[cryptography]==3.3.0
python-multipart==0.0.20
httpx==0.28.1
pytest==8.3.4
pytest-asyncio==0.25.0
